packs/pyproject.toml
[project]
name = "server-tools"
version = "0.1.0"
description = "Python scripts for managing the Solar System Bedrock server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "nbtlib>=2.0.4",
    "Pillow>=10.0.0",
    "textual>=0.80.0",
]
